MakinoPhysicsEngineを搭載。それに伴って必要だった修正や改良を行いました。

+++ MakinoProject00/MakinoProject00/Model.cpp
@@ -41,5 +41,5 @@ CSkeletalModel::CSkeletalModel(CGameObject* owner, GraphicsLayer layer, const st
 
 // Pre drawing processing
 void CSkeletalModel::PreDraw() {
-    m_model.StepAnim((float)CAppClock::GetMain().GetDeltaTime());
+    m_model.StepAnim(CAppClock::GetMain().GetDeltaTime());
 }
